Q1 - Q4 Answers
Q1
The ‘onLogout’ function receives a ‘Player object’ that is unsafe to send as a parameter to the ‘addEvent’ call.

Sending an Id, which is a plain number and therefore safe, will give ‘releaseStorage’ enough information to
reconstruct the Player object set their storage value.

Q2
The variable ‘result’ wasn’t defined anywhere, so the call to ‘getString()’ would’ve failed
and ‘guildName’ would ultimately be nil

Additionally, no iteration was being done on the query result (which should be a table),
which also makes it impossible to properly print the guilds’ names.

Renamed ‘resultId’ to ‘guildsFound’ and added a for loop to iterate through the guilds that
matched the criteria and print their names.

Q3
The function removes a Player from its current party.

A Player can only be part of a single Party at most so we can derive the Party
information just from the Player we wish to remove from it, no need to pass extra
Player data in the parameters.

There were a lot of repeated calls to ‘Player(membername)’ which is inefficient.
These calls have been condensed into the single ‘player’ variable.

Added an early out in the event that the Player doesn’t currently belong to any party.

Q4
Creating an empty Player through 'new Player(nullptr)' will cause a memory
leak because the new object is not being deleted manually afterwards.

Creating it without the 'new' keyword will ensure the Player object will
be cleaned up as soon as the function is done.